requirements.txt
beautifulsoup4
pillow
pyyaml
requests
